action.js
Actions (sometimes called action creators) are functions that you call to send data into the system. In Nuclear, any function that calls reactor.dispatch(actionType: string, payload: any) is categorized as an action.

For our example, we'll start by creating an action to fetch products from a server and another action to add a product to the user's shopping cart.

In order to correctly reference actions throughout the system, we'll create an actionTypes.js file, which is simply a collection of constants. We're using React's keyMirror utility to create the constants, but that's just a convenience — you can create action types in any way you'd like. They're not even required to be in a separate file, though that's certainly recommended