Add files via upload
Is the 'Find devices and content' setting turned to On?

Windows 8 and 8.1 have sharing settings that must be set to ON. The steps to do this differ slightly depending on operating system version.

On Windows 8.1 (most users), you can do the following:

Navigate to Settings. To do this, move your mouse to the far right of the screen to show the menu bar. Settings will be at the bottom.
Select Change PC Settings (at the very bottom of the Settings menu).
Select Network.
Select your connected Wi-Fi network (next to Wi-Fi bar image).
Under Find Devices and Content, make sure it is set to On.
If you are running Windows 8, you can do the following:

Navigate to your Desktop. (Your Desktop will be one of the default tiles on your homescreen).
Click on your Wi-Fi settings on the bottom of your desktop bar (at the very bottom, right hand side of your Desktop).
Right click on your connected Wi-Fi network.
Select Turn sharing on or off.
Make sure sharing is set to On.
Do you have